Refactor(models): reorganize files.
