+++ api_nestbook/app/models/editorial.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel, Field
+from datetime import date
+from typing import Optional
+
+class EditorialBase(BaseModel):
+    name_editorial: str
+
+class EditorialDb(EditorialBase):
+    id: int
